fix: add support to changed special character handling (#4)
Newer versions of tmux (observed on 3.4, on Ubuntu 24.04) in comparison
to the older one (observed on 3.2a, on Ubuntu 22.04) changed behavior
in how special characters within the option values are returned
by the command `tmux show-option -gqv`. Now, it adds a leading backslash
to each special character.
This breaks the expansion of environment variables that is done by the
function `resurrect_dir()` (in #/scripts/helpers.sh).
This commit adds support for this new representation of special
characters, in addition to the older ones.

Signed-off-by: Arseniy Aharonov <aaharonov@nvidia.com>
authored-by: Arseniy Aharonov <aaharonov@nvidia.com>